Refactoring get Method in Model
