+++ graphql/database/base/RedisModel.js
@@ -1,6 +1,4 @@
 import Redis from 'ioredis'
-import { isObject } from 'lodash/lang'
-import { get } from 'lodash/object'
 import Model from './Model'
 import {
   DatabaseConditionalCheckFailedException,
@@ -8,21 +6,13 @@ import {
   FieldDoesNotExistException,
   UnauthorizedQueryException,
 } from '../../utils/exceptions'
-import logger from '../../utils/logger'
 
 const client = new Redis(process.env.UPSTASH_HOST)
 
 class RedisModel extends Model {
-  static async getInternal(keys) {
-    // RedisModel only understands the concept of hash keys;
-    // get should not be called with a range key.
-    if (keys.length > 1) {
-      logger.warn(
-        `RedisModel get should only be called with a hashKey, received ${keys.toString()}`
-      )
-    }
-    const key = this.getRedisKey(keys[0])
-    const item = await client.hgetall(key)
+  static async getInternal(key) {
+    const redisKey = this.getRedisKey(key)
+    const item = await client.hgetall(redisKey)
     if (Object.keys(item).length === 0) {
       throw new DatabaseItemDoesNotExistException()
     }
@@ -30,13 +20,6 @@ class RedisModel extends Model {
   }
 
   static async getBatchInternal(keys) {
-    keys.forEach(key => {
-      if (isObject(key)) {
-        throw new Error(
-          `RedisModel getBatch only receives a list of hash keys, received ${keys.toString()}`
-        )
-      }
-    })
     const items = keys.map(key => this.getInternal(key))
     return Promise.all(items)
   }
@@ -59,8 +42,7 @@ class RedisModel extends Model {
   }
 
   static async updateInternal(item) {
-    const self = this
-    const redisKey = this.getRedisKey(get(item, self.hashKey))
+    const redisKey = this.getRedisKey(item.id)
     const existingEntry = await client.hgetall(redisKey)
     if (Object.keys(existingEntry).length === 0) {
       throw new DatabaseConditionalCheckFailedException()
